F5,F2,F3

+++ MyFood/Models/Form3ViewModel.cs
@@ -54,13 +54,13 @@ namespace MyFood.Models
         public TimeSpan? return_time { get; set; }
 
         [Display(Name = "عدد الكيلوات أثناء الخروج")]
-        public byte? kilos_exit_time { get; set; }
+        public int? kilos_exit_time { get; set; }
 
         [Display(Name = "عدد الكيلوات أثناء الوصول")]
-        public byte? kilos_arrival_time { get; set; }
+        public int? kilos_arrival_time { get; set; }
 
         [Display(Name = "عدد الكيلوات أثناء الرجوع")]
-        public byte? kilos_return_time { get; set; }
+        public int? kilos_return_time { get; set; }
 
         [StringLength(10)]
         public string note { get; set; }
@@ -74,7 +74,7 @@ namespace MyFood.Models
         [ForeignKey("Order")]
         public long order_id { get; set; }
         public Order Order { get; set; }
-        public byte? food_type_id { get; set; }
+        public int? food_type_id { get; set; }
         public FoodType FoodType { get; set; }
 
         [ForeignKey("teamLeaderId")]
